requirements.txt
numpy~=1.19.2
pillow~=8.1.2
pyqt~=5.9.2
pyqtgraph~=0.11.0
Qt~=5.9.7